feat: Create `MaterialSystem` for graphics calls
The `MaterialSystem` is a layer of abstraction to the OpenGL API. What
this structure seeks is to provide a reference counted window into
OpenGL, providing simple functions and structures that can be called to
quickly create and configure objects for rendering. The overall goal is
to make calls to the underlying rendering system significantly easier
with little to no hassle. This is the current early stage, but as this
renderer grows larger, the `MaterialSystem` will provide an abstraction
that will greatly improve readability and scalability as time goes on.

More can be added to this class in the future, as the current state of
the system is just to provide some abstraction. In the future, this
could handle much more than just that.